Section 4 - Complete

+++ HR.LeaveManagement.Application/Features/LeaveAllocations/Handlers/Commands/Delete_LeaveAllocationCommandHandler.cs
@@ -1,6 +1,10 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.DTOs.LeaveAllocation.Validators;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveAllocations.Requests.Commands;
-using HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Responses;
+using HR.LeaveManagement.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -10,24 +14,46 @@ using System.Threading.Tasks;
 
 namespace HR.LeaveManagement.Application.Features.LeaveAllocations.Handlers.Commands
 {
-    public class Delete_LeaveAllocationCommandHandler : IRequestHandler<Delete_LeaveAllocationCommand>
+    public class Delete_LeaveAllocationCommandHandler : IRequestHandler<Delete_LeaveAllocationCommand, BaseCommandResponse>
     {
         private readonly ILeaveAllocationRepository _leaveAllocationRepository;
+        private readonly ILeaveTypeRepository _leaveTypeRepository;
         private readonly IMapper _mapper;
 
-        public Delete_LeaveAllocationCommandHandler(ILeaveAllocationRepository leaveAllocationRepository, IMapper mapper)
+        public Delete_LeaveAllocationCommandHandler(ILeaveAllocationRepository leaveAllocationRepository, ILeaveTypeRepository leaveTypeRepository, IMapper mapper)
         {
             _leaveAllocationRepository = leaveAllocationRepository;
+            _leaveTypeRepository = leaveTypeRepository;
             _mapper = mapper;
         }
 
-        public async Task<Unit> Handle(Delete_LeaveAllocationCommand request, CancellationToken cancellationToken)
+        public async Task<BaseCommandResponse> Handle(Delete_LeaveAllocationCommand request, CancellationToken cancellationToken)
         {
-            var leaveAllocation = await _leaveAllocationRepository.GetAsync(request.Id);
+            var response = new BaseCommandResponse();
+            var validator = new Delete_LeaveAllocationDTOValidator(_leaveTypeRepository);
+            var validationResult = await validator.ValidateAsync(request.LeaveAllocationDTO);
+
+            if (!validationResult.IsValid)
+            {
+                response.Success = false;
+                response.Message = "Failed to delete record!";
+                response.Errors = validationResult.Errors.Select(e => e.ErrorMessage).ToList();
+            }
+
+            var leaveAllocation = await _leaveAllocationRepository.GetAsync(request.LeaveAllocationDTO.Id);
+
+            if (leaveAllocation == null)
+            {
+                throw new NotFoundException(nameof(LeaveAllocation), request.LeaveAllocationDTO.Id);
+            }
 
             await _leaveAllocationRepository.DeleteAsync(leaveAllocation);
 
-            return Unit.Value;
+            response.Success = true;
+            response.Message = "Record deleted.";
+            response.Id = leaveAllocation.Id;
+
+            return response;
         }
     }
 }
