+++ HR.LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/Delete_LeaveRequestCommandHandler.cs
@@ -1,6 +1,10 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.DTOs.LeaveRequest.Validators;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
-using HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Responses;
+using HR.LeaveManagement.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -10,24 +14,46 @@ using System.Threading.Tasks;
 
 namespace HR.LeaveManagement.Application.Features.LeaveRequests.Handlers.Commands
 {
-    public class Delete_LeaveRequestCommandHandler : IRequestHandler<Delete_LeaveRequestCommand>
+    public class Delete_LeaveRequestCommandHandler : IRequestHandler<Delete_LeaveRequestCommand, BaseCommandResponse>
     {
         private readonly ILeaveRequestRepository _leaveRequestRepository;
+        private readonly ILeaveTypeRepository _leaveTypeRepository;
         private readonly IMapper _mapper;
 
-        public Delete_LeaveRequestCommandHandler(ILeaveRequestRepository leaveRequestRepository, IMapper mapper)
+        public Delete_LeaveRequestCommandHandler(ILeaveRequestRepository leaveRequestRepository, ILeaveTypeRepository leaveTypeRepository, IMapper mapper)
         {
             _leaveRequestRepository = leaveRequestRepository;
+            _leaveTypeRepository = leaveTypeRepository;
             _mapper = mapper;
         }
 
-        public async Task<Unit> Handle(Delete_LeaveRequestCommand request, CancellationToken cancellationToken)
+        public async Task<BaseCommandResponse> Handle(Delete_LeaveRequestCommand request, CancellationToken cancellationToken)
         {
+            var response = new BaseCommandResponse();
+            var validator = new Delete_LeaveRequestDTOValidator(_leaveTypeRepository);
+            var validationResult = await validator.ValidateAsync(request.LeaveRequestDTO);
+
+            if (!validationResult.IsValid)
+            {
+                response.Success = false;
+                response.Message = "Failed to delete record!";
+                response.Errors = validationResult.Errors.Select(e => e.ErrorMessage).ToList();
+            }
+
             var leaveRequest = await _leaveRequestRepository.GetAsync(request.Id);
+
+            if (leaveRequest == null)
+            {
+                throw new NotFoundException(nameof(LeaveRequest), request.Id);
+            }
 
             await _leaveRequestRepository.DeleteAsync(leaveRequest);
 
-            return Unit.Value;
+            response.Success = true;
+            response.Message = "Record deleted.";
+            response.Id = leaveRequest.Id;
+
+            return response;
         }
     }
 }
