+++ HR.LeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/Delete_LeaveTypeCommandHandler.cs
@@ -1,7 +1,11 @@
 ﻿using AutoMapper;
 using HR.LeaveManagement.Application.DTOs.LeaveType;
+using HR.LeaveManagement.Application.DTOs.LeaveType.Validators;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
-using HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Responses;
+using HR.LeaveManagement.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -11,7 +15,7 @@ using System.Threading.Tasks;
 
 namespace HR.LeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
 {
-    public class Delete_LeaveTypeCommandHandler : IRequestHandler<Delete_LeaveTypeCommand>
+    public class Delete_LeaveTypeCommandHandler : IRequestHandler<Delete_LeaveTypeCommand, BaseCommandResponse>
     {
         private readonly ILeaveTypeRepository _leaveTypeRepository;
         private readonly IMapper _mapper;
@@ -22,13 +26,33 @@ namespace HR.LeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
             _mapper = mapper;
         }
 
-        public async Task<Unit> Handle(Delete_LeaveTypeCommand request, CancellationToken cancellationToken)
+        public async Task<BaseCommandResponse> Handle(Delete_LeaveTypeCommand request, CancellationToken cancellationToken)
         {
-            var leaveType = await _leaveTypeRepository.GetAsync(request.Id);
+            var response = new BaseCommandResponse();
+            var validator = new Delete_LeaveTypeDTOValidator();
+            var validationResult = await validator.ValidateAsync(request.LeaveTypeDTO);
+
+            if (!validationResult.IsValid)
+            {
+                response.Success = false;
+                response.Message = "Failed to delete record!";
+                response.Errors = validationResult.Errors.Select(e => e.ErrorMessage).ToList();
+            }
+
+            var leaveType = await _leaveTypeRepository.GetAsync(request.LeaveTypeDTO.Id);
+
+            if (leaveType == null)
+            {
+                throw new NotFoundException(nameof(LeaveType), request.LeaveTypeDTO);
+            }
 
             await _leaveTypeRepository.DeleteAsync(leaveType);
 
-            return Unit.Value;
+            response.Success = true;
+            response.Message = "Record deleted.";
+            response.Id = leaveType.Id;
+
+            return response;
         }
     }
 }
